@@ -1,8 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdk 32
 
     defaultConfig {
         applicationId "com.stibla.threedskitracker"
@@ -18,7 +17,7 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:+'
-    implementation 'com.google.android.gms:play-services-maps:+'
-    implementation 'com.android.support:appcompat-v7:21.0.3'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.gms:play-services-maps:18.0.2'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
 }
